Test
This looks like its also going to be a 2 parter, exploring environments and detailed denizens.

Here is my issue; from the diverse monsters we could introduce to needed player options to support this style of adventure, this kind of article would be excellent to write and entire book around. This could easily extend to 30-50 pages and make an excellent small module.